add separate methods for each job, view, configFile type.
