+++ job-dsl-core/src/main/groovy/javaposse/jobdsl/dsl/ConfigFileType.groovy
@@ -1,12 +1,6 @@
 package javaposse.jobdsl.dsl
 
 enum ConfigFileType {
-    Custom(ConfigFile),
-    MavenSettings(ConfigFile)
-
-    final Class<? extends ConfigFile> configFileClass
-
-    ConfigFileType(Class<? extends ConfigFile> configFileClass) {
-        this.configFileClass = configFileClass
-    }
+    Custom,
+    MavenSettings
 }
