+++ job-dsl-core/src/main/groovy/javaposse/jobdsl/dsl/helpers/triggers/TriggerContext.groovy
@@ -5,19 +5,16 @@ import javaposse.jobdsl.dsl.Context
 import javaposse.jobdsl.dsl.ContextHelper
 import javaposse.jobdsl.dsl.DslContext
 import javaposse.jobdsl.dsl.JobManagement
-import javaposse.jobdsl.dsl.JobType
 import javaposse.jobdsl.dsl.WithXmlAction
 import javaposse.jobdsl.dsl.helpers.triggers.GerritContext.GerritSpec
 
 class TriggerContext implements Context {
-    private final List<WithXmlAction> withXmlActions
-    private final JobType jobType
-    private final JobManagement jobManagement
+    protected final List<WithXmlAction> withXmlActions
+    protected final JobManagement jobManagement
     final List<Node> triggerNodes = []
 
-    TriggerContext(List<WithXmlAction> withXmlActions, JobType jobType, JobManagement jobManagement) {
+    TriggerContext(List<WithXmlAction> withXmlActions, JobManagement jobManagement) {
         this.withXmlActions = withXmlActions
-        this.jobType = jobType
         this.jobManagement = jobManagement
     }
 
@@ -278,20 +275,5 @@ class TriggerContext implements Context {
         }
 
         triggerNodes << gerritNode
-    }
-
-    /**
-     * If set to <code>true</code>, Jenkins will parse the POMs of this project, and see if any of its snapshot
-     * dependencies are built on this Jenkins as well. If so, Jenkins will set up build dependency relationship so that
-     * whenever the dependency job is built and a new SNAPSHOT jar is created, Jenkins will schedule a build of this
-     * project. Defaults to <code>false</code>.
-     * @param checkSnapshotDependencies set to <code>true</code> to check snapshot dependencies
-     */
-    void snapshotDependencies(boolean checkSnapshotDependencies) {
-        Preconditions.checkState jobType == JobType.Maven, 'snapshotDependencies can only be applied for Maven jobs'
-        withXmlActions << WithXmlAction.create {
-            it.children().removeAll { it instanceof Node && it.name() == 'ignoreUpstremChanges' }
-            it.appendNode 'ignoreUpstremChanges', !checkSnapshotDependencies
-        }
     }
 }
