+++ job-dsl-core/src/main/groovy/javaposse/jobdsl/dsl/helpers/wrapper/WrapperContext.groovy
@@ -6,17 +6,14 @@ import javaposse.jobdsl.dsl.Context
 import javaposse.jobdsl.dsl.ContextHelper
 import javaposse.jobdsl.dsl.DslContext
 import javaposse.jobdsl.dsl.JobManagement
-import javaposse.jobdsl.dsl.JobType
 import javaposse.jobdsl.dsl.WithXmlAction
 
 class WrapperContext implements Context {
     List<Node> wrapperNodes = []
-    JobType type
     JobManagement jobManagement
 
-    WrapperContext(JobType jobType, JobManagement jobManagement) {
+    WrapperContext(JobManagement jobManagement) {
         this.jobManagement = jobManagement
-        this.type = jobType
     }
 
     void timestamps() {
@@ -608,47 +605,6 @@ class WrapperContext implements Context {
                     }
                 }
             }
-        }
-    }
-
-    /**
-     * <p>Configures a release using the m2release plugin.</p>
-     * <p>By default the following values are applied. If an instance of a
-     * closure is provided, the values from the closure will take effect.</p>
-     * <pre>
-     * {@code
-     * <buildWrappers>
-     *     <org.jvnet.hudson.plugins.m2release.M2ReleaseBuildWrapper>
-     *         <scmUserEnvVar></scmUserEnvVar>
-     *         <scmPasswordEnvVar></scmPasswordEnvVar>
-     *         <releaseEnvVar>IS_M2RELEASEBUILD</releaseEnvVar>
-     *         <releaseGoals>-Dresume=false release:prepare release:perform</releaseGoals>
-     *         <dryRunGoals>-Dresume=false -DdryRun=true release:prepare</dryRunGoals>
-     *         <selectCustomScmCommentPrefix>false</selectCustomScmCommentPrefix>
-     *         <selectAppendHudsonUsername>false</selectAppendHudsonUsername>
-     *         <selectScmCredentials>false</selectScmCredentials>
-     *         <numberOfReleaseBuildsToKeep>1</numberOfReleaseBuildsToKeep>
-     *     </org.jvnet.hudson.plugins.m2release.M2ReleaseBuildWrapper>
-     * </buildWrappers>
-     *}
-     * </pre>
-     */
-    void mavenRelease(@DslContext(MavenReleaseContext) Closure releaseClosure = null) {
-        Preconditions.checkState type == JobType.Maven, 'mavenRelease can only be applied for Maven jobs'
-
-        MavenReleaseContext context = new MavenReleaseContext()
-        ContextHelper.executeInContext(releaseClosure, context)
-
-        wrapperNodes << new NodeBuilder().'org.jvnet.hudson.plugins.m2release.M2ReleaseBuildWrapper' {
-            scmUserEnvVar context.scmUserEnvVar
-            scmPasswordEnvVar context.scmPasswordEnvVar
-            releaseEnvVar context.releaseEnvVar
-            releaseGoals context.releaseGoals
-            dryRunGoals context.dryRunGoals
-            selectCustomScmCommentPrefix context.selectCustomScmCommentPrefix
-            selectAppendHudsonUsername context.selectAppendJenkinsUsername
-            selectScmCredentials context.selectScmCredentials
-            numberOfReleaseBuildsToKeep context.numberOfReleaseBuildsToKeep
         }
     }
