+++ job-dsl-core/src/test/groovy/javaposse/jobdsl/dsl/ConfigFileSpec.groovy
@@ -3,14 +3,12 @@ package javaposse.jobdsl.dsl
 import spock.lang.Specification
 
 class ConfigFileSpec extends Specification {
+
     private final JobManagement jobManagement = Mock(JobManagement)
     private final ConfigFile configFile = new ConfigFile(ConfigFileType.Custom, jobManagement)
 
     def 'get type'() {
-        when:
-        ConfigFile configFile = new ConfigFile(ConfigFileType.Custom, jobManagement)
-
-        then:
+        expect:
         configFile.type == ConfigFileType.Custom
     }
